pyproject.toml
[tool.poetry]
name = "stc-backend"
version = "1.0.0"
description = "Save The Children - FastAPI Backend"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.8.0"
pydantic-settings = "^2.1.0"
motor = "^3.3.2"
pymongo = "^4.6.0"
python-dotenv = "^1.0.0"
PyJWT = "^2.8.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.1"
python-multipart = "^0.0.6"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
aiofiles = "^23.2.1"
azure-storage-blob = "^12.19.0"
langchain = "^0.2.0"
langchain-groq = "^0.1.2"
langchain-google-genai = "^1.0.0"
pinecone-client = "^5.0.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
